fix: implement request submission hook and update form management
