+++ src/pages/holidays/RequestEnjoyment/index.tsx
@@ -1,13 +1,10 @@
 import { useRef, useState } from "react";
-import { useNavigate } from "react-router-dom";
+
 import { FormikProps } from "formik";
 
-import { postHumanResourceRequest } from "@src/services/humanResourcesRequest/postHumanResourceRequest";
-import { IRequestBody } from "@src/services/humanResourcesRequest/postHumanResourceRequest/types";
 import { SendRequestModal } from "@components/modals/SendRequestModal";
 import { RequestInfoModal } from "@components/modals/RequestInfoModal";
-import { useAppContext } from "@context/AppContext/useAppContext";
-import { formatDate } from "@utils/date";
+
 import { useErrorFlag } from "@hooks/useErrorFlag";
 
 import { IGeneralInformationEntry } from "./forms/GeneralInformationForm/types";
@@ -15,6 +12,7 @@ import { RequestEnjoymentUI } from "./interface";
 import { requestEnjoymentSteps } from "./config/assisted.config";
 import { holidaysNavConfig } from "../config/nav.config";
 import { ModalState } from "./types";
+import { useRequestSubmission } from "@src/hooks/usePostHumanResourceResquest";
 
 function useFormManagement() {
   const [formValues, setFormValues] = useState<IGeneralInformationEntry>({
@@ -71,84 +69,6 @@ function useModalManagement() {
   };
 }
 
-function useRequestSubmission(formValues: IGeneralInformationEntry) {
-  const [requestNum, setRequestNum] = useState("");
-  const [staffName, setStaffName] = useState<string | null>(null);
-  const { selectedEmployee, requestsHolidays, setRequestsHolidays } =
-    useAppContext();
-  const navigate = useNavigate();
-
-  const [showErrorFlag, setShowErrorFlag] = useState(false);
-  const [errorMessage, setErrorMessage] = useState("");
-
-  const submitRequest = async () => {
-    try {
-      const humanResourceRequestData = JSON.stringify({
-        daysOff: formValues.daysOff,
-        startDate: formatDate(formValues.startDate),
-        contract: formValues.contract,
-      });
-
-      const userCodeInCharge = "User 1";
-      const userNameInCharge = "Johan Daniel Garcia Nova";
-
-      const requestBody: IRequestBody = {
-        employeeId: selectedEmployee.employeeId,
-        humanResourceRequestData: humanResourceRequestData,
-        humanResourceRequestDate: new Date().toISOString(),
-        humanResourceRequestDescription: formValues.observations || "",
-        humanResourceRequestStatus: "in_progress",
-        humanResourceRequestType: "vacations",
-        userCodeInCharge,
-        userNameInCharge,
-      };
-
-      if (userCodeInCharge && userNameInCharge) {
-        setStaffName(userNameInCharge);
-      } else {
-        setStaffName(null);
-      }
-
-      const response = await postHumanResourceRequest(requestBody);
-
-      if (response?.humanResourceRequestId) {
-        setRequestNum(response.humanResourceRequestNumber);
-        setRequestsHolidays([...requestsHolidays, requestBody]);
-        return true;
-      }
-      return false;
-    } catch (error) {
-      console.error("Error sending request:", error);
-      setErrorMessage(
-        "Error al enviar la solicitud de vacaciones. Intente nuevamente.",
-      );
-      setShowErrorFlag(true);
-      return false;
-    }
-  };
-
-  const navigateAfterSubmission = () => {
-    navigate("/holidays", {
-      state: {
-        showFlag: true,
-        flagTitle: "Solicitud enviada",
-        flagMessage: "La solicitud de certificación fue enviada exitosamente.",
-        isSuccess: true,
-      },
-    });
-  };
-
-  return {
-    requestId: requestNum,
-    staffName,
-    submitRequest,
-    navigateAfterSubmission,
-    showErrorFlag,
-    errorMessage,
-    setShowErrorFlag,
-  };
-}
-
 function RequestEnjoyment() {
   const [currentStep, setCurrentStep] = useState(1);
   const {
@@ -173,7 +93,7 @@ function RequestEnjoyment() {
     showErrorFlag,
     errorMessage,
     setShowErrorFlag,
-  } = useRequestSubmission(formValues);
+  } = useRequestSubmission(formValues, "vacations");
 
   useErrorFlag(showErrorFlag, errorMessage, "Error", false, 10000);
