+++ src/types/humanResourcesRequest.types.ts
@@ -67,4 +67,12 @@ export enum ETaskStatus {
   Executed = "Ejecutada",
 }
 
+export interface IGeneralInformationEntry {
+  id: string;
+  daysOff: string;
+  startDate: string;
+  contract: string;
+  observations: string;
+}
+
 export type HumanResourceRequests = HumanResourceRequest[];
